Updated active skill selection color
As seen in this video (https://www.youtube.com/watch?v=iigm6t6PzP8&list=UU2MlJgYpvoJ0Cz-Jy5P42uQ), once a skill is selected and made active, the active color is blue, not green!

However, I can't remember if it's blue or green currently in the Alpha. So disregard if this is the case.